[6] rename folder and files
